Make the Mach-O tree clippy-clean under edition 2024
The imported code was written for edition 2021. Collapse the nested ifs
into let chains, annotate transmutes, use is_multiple_of and checked
division, drop redundant casts and borrows, add the Default impls
clippy asks for, remove an unused type parameter and a recursion-only
parameter, move copy_buf out of a module named after its file, and
simplify a filter that clippy proves is equivalent to obj.is_alive (the
only_alive flag was already implied by the second conjunct).

+++ src/macho/arch/x86_64.rs
@@ -170,7 +170,7 @@ impl Arch for X86_64 {
             let is_subtracted = i > 0 && rels[i - 1].r_type() == X86_64_RELOC_SUBTRACTOR;
 
             let (target, addend) = if r.is_extern() {
-                (RelocTarget::Sym(r.r_symbolnum() as u32), addend)
+                (RelocTarget::Sym(r.r_symbolnum()), addend)
             } else {
                 let addr = if r.is_pcrel() {
                     (hdr.addr + r.r_address as u64 + 4).wrapping_add_signed(addend)
